[Chore] Separate out `vllm.utils.platform_utils.py` (#27374)
