+++ vllm/lora/models.py
@@ -31,8 +31,8 @@ from vllm.model_executor.models import SupportsLoRA, supports_multimodal
 from vllm.model_executor.models.interfaces import is_pooling_model
 from vllm.model_executor.models.module_mapping import MultiModelKeys
 from vllm.model_executor.models.utils import PPMissingLayer, WeightsMapper
-from vllm.utils import is_pin_memory_available
 from vllm.utils.cache import LRUCache
+from vllm.utils.platform_utils import is_pin_memory_available
 
 logger = init_logger(__name__)
